Moved MOMF to new file and added/modified files suggested in second PR review
